pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gonka-wallet"
version = "0.1.0"
description = "Gonka wallet library for the Gonka chain"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "raferti"},
]
keywords = ["cosmos", "blockchain", "wallet", "gonka"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mnemonic",
    "bip_utils",
    "bech32",
    "pycryptodome",
    "httpx",
    "betterproto",
    "ecdsa",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.urls]
Repository = "https://github.com/raferti/gonka-wallet"

[tool.setuptools.packages.find]
include = ["gonka_wallet*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
